Conclusao apis pedidos

+++ src/main/java/br/com/rdevs/ecommerce/pedido/model/dto/StatusPedidoDTO.java
@@ -8,6 +8,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class StatusPedidoDTO {
-    private Integer cdStatusPedido;
+    private Long cdStatusPedido;
     private String dsStatusPedido;
 }
